bug fix, energy long double back to double

+++ nested_sampling/src/cv.c
@@ -3,9 +3,9 @@
 #include <stdio.h>
 double X_imp(int i, double K, double P);
 void compute_dos(long double* gl, int N, double P, double K);
-void renorm_energies(long double* El, int N, long double Emin);
-double heat_capacity(long double* El, long double* gl, int N, double T, double ndof);
-void heat_capacity_loop(long double* El, long double* gl, double* Cvl, int N, double Tmin, double Tmax, int nT, double ndof);
+void renorm_energies(double* El, int N, double Emin);
+double heat_capacity(double* El, long double* gl, int N, double T, double ndof);
+void heat_capacity_loop(double* El, long double* gl, double* Cvl, int N, double Tmin, double Tmax, int nT, double ndof);
 
 double X_imp(int i, double K, double P)
 {
@@ -70,7 +70,7 @@ void compute_dos_imp(long double* gl, int N, double P, double K)
 }
 
 ////////////renormalise energies wrt ground state/////////////////
-void renorm_energies(long double* El, int N, long double Emin)
+void renorm_energies(double* El, int N, double Emin)
 {
   int i;
   for(i=0;i<N;++i)
@@ -80,7 +80,7 @@ void renorm_energies(long double* El, int N, long double Emin)
 }
 
 ////////////////////////////caclulate heat capacity for a single T////////////////////////
-double heat_capacity(long double* El, long double* gl, int N, double T, double ndof)
+double heat_capacity(double* El, long double* gl, int N, double T, double ndof)
 {
   //K is the number of replicas, beta the reduced temperature and E is the array of energies 
   int i;
@@ -107,7 +107,7 @@ double heat_capacity(long double* El, long double* gl, int N, double T, double n
 }
 
 //////////////////////////////calculate heat capacity over a set of Ts/////////////////////
-void heat_capacity_loop(long double* El, long double* gl, double* Cvl, int N, double Tmin, double Tmax, int nT, double ndof)
+void heat_capacity_loop(double* El, long double* gl, double* Cvl, int N, double Tmin, double Tmax, int nT, double ndof)
 {
   //Cvl is a 0's array of size N (same size as El)
   int i;
